--- imageHandler.py
@@ -1,61 +1,0 @@
-import requests
-import io
-from PIL import Image
-import json
-from resources.env import DEBUG, VERBOSE
-from resources.textColors import redText, greenText, blueText
-
-class ImageObject :
-    def __init__(self, src="", label="", filename="") :
-        self.src = src
-        self.label=label
-        self.filename=filename
-        if (src == "" or label == "") :
-            raise Exception("Invalid Image Object Initializers")
-    def getSrc(self) :
-        if (self.src) :
-            return self.src
-        else :
-            if (DEBUG) : print("No image Src")
-    def getLabel(self) :
-        if (self.label) :
-            return self.label
-        else :
-            if (DEBUG) : print("No image label")
-    def getFilename(self) :
-        if (self.filename) :
-            return self.filename
-        else :
-            if (DEBUG) : print("No image filename")
-    def setFilename(self, filename) :
-        self.filename = filename
-    def imageJSON(self) :
-        data = {
-            "label" : str(self.label),
-            "src" : str(self.src),
-            "filename" : str(self.filename)
-        }
-        return data
-    def downloadImage(self, downloadPath) :
-        try :
-            image_content = requests.get(self.src).content
-            if (DEBUG) : print(greenText("Successful Request: " + self.src))
-            image_file = io.BytesIO(image_content)
-            image = Image.open(image_file)
-            file_path = downloadPath + self.filename + ".jpg"
-
-            with open(file_path, "wb") as f:
-                image.save(f, "JPEG")
-
-            if (DEBUG) : print(greenText("Saved image: " + file_path))
-        except Exception as e:
-            if (DEBUG) : print(redText('FAILED -' + str(e)))
-
-
-
-def outputJSON(outPath, jsonFilename, jsonData) :
-    jsonObject = json.dumps(jsonData, indent=3)
-    filePath = outPath + jsonFilename + ".json"
-    with open(filePath, "w") as outfile :
-        json.dump(jsonObject, outfile, indent=3)
-        if (DEBUG) : print(greenText("Output to JSON success"))
