+++ resources/env.py
@@ -1,10 +1,10 @@
 # Environment
-PATH = "C:\Users\evan9\OneDrive\Desktop\school\Applied Machine Learning\project2\scraper"
+PATH = "C:\\Users\\evan9\\OneDrive\\Desktop\\school\\Applied Machine Learning\\project2\\scraper\\"
 IMAGE_DIR = "images\\"
 DATA_DIR = "data\\"
 MAX_IMAGES = 6
-CLASSES = ["cat"]
-BINARY_LOCATION = "C:\\Program Files\\Mozilla Firefox\\firefox.exe"
+CLASSES = ["cat", "dog"]
+BROWSER_LOCATION = "C:\\Program Files\\Mozilla Firefox\\firefox.exe"
 DRIVER = "resources\\driver.exe"
 #Output
 DEBUG = True #Do you want debug messages?
